+++ packages/core-cairo/src/erc20.ts
@@ -10,7 +10,7 @@ import { defineComponents } from './utils/define-components';
 import { defaults as commonDefaults } from './common-options';
 import { printContract } from './print';
 import { externalTrait } from './external-trait';
-import { toShortString } from './utils/convert-strings';
+import { toStringLiteral } from './utils/convert-strings';
 
 export const defaults: Required<ERC20Options> = {
   name: 'MyToken',
@@ -57,8 +57,8 @@ export function buildERC20(opts: ERC20Options): Contract {
 
   const allOpts = withDefaults(opts);
 
-  addBase(c, toShortString(allOpts.name, 'name'), toShortString(allOpts.symbol, 'symbol'));
-  addERC20ImplAndCamelOnlyImpl(c, allOpts.pausable);
+  addBase(c, toStringLiteral(allOpts.name), toStringLiteral(allOpts.symbol));
+  addERC20MixinOrImpls(c, allOpts.pausable);
 
   if (allOpts.premint) {
     addPremint(c, allOpts.premint);
@@ -93,9 +93,14 @@ function addERC20Interface(c: ContractBuilder) {
   c.addStandaloneImport('openzeppelin::token::erc20::interface');
 }
 
-function addERC20ImplAndCamelOnlyImpl(c: ContractBuilder, pausable: boolean) {
+function addERC20MixinOrImpls(c: ContractBuilder, pausable: boolean) {
   if (pausable) {
     addERC20Interface(c);
+
+    c.addImplToComponent(components.ERC20Component, {
+      name: 'ERC20MetadataImpl',
+      value: 'ERC20Component::ERC20MetadataImpl<ContractState>',
+    });
 
     const ERC20Impl: BaseImplementedTrait = {
       name: 'ERC20Impl',
@@ -118,17 +123,14 @@ function addERC20ImplAndCamelOnlyImpl(c: ContractBuilder, pausable: boolean) {
         'abi(embed_v0)'
       ],
     }
+    // Camel case versions of the functions above. Pausable is already set above.
     c.addFunction(ERC20CamelOnlyImpl, functions.totalSupply);
     c.addFunction(ERC20CamelOnlyImpl, functions.balanceOf);
-    setPausable(c, ERC20CamelOnlyImpl, functions.transferFrom);
+    c.addFunction(ERC20CamelOnlyImpl, functions.transferFrom);
   } else {
     c.addImplToComponent(components.ERC20Component, {
-      name: 'ERC20Impl',
-      value: 'ERC20Component::ERC20Impl<ContractState>',
-    });
-    c.addImplToComponent(components.ERC20Component,     {
-      name: 'ERC20CamelOnlyImpl',
-      value: 'ERC20Component::ERC20CamelOnlyImpl<ContractState>',
+      name: 'ERC20MixinImpl',
+      value: 'ERC20Component::ERC20MixinImpl<ContractState>',
     });
   }
 }
@@ -223,12 +225,7 @@ const components = defineComponents( {
       name: 'ERC20Event',
       type: 'ERC20Component::Event',
     },
-    impls: [
-      {
-        name: 'ERC20MetadataImpl',
-        value: 'ERC20Component::ERC20MetadataImpl<ContractState>',
-      },    
-    ],
+    impls: [],
     internalImpl: {
       name: 'ERC20InternalImpl',
       value: 'ERC20Component::InternalImpl<ContractState>',