+++ packages/core-cairo/src/utils/convert-strings.test.ts
@@ -1,7 +1,7 @@
 import test from 'ava';
 
-import { toIdentifier, toShortString } from './convert-strings';
-import type { OptionsError } from '../error';
+import { toIdentifier, toStringLiteral } from './convert-strings';
+import { OptionsError } from '../error';
 
 test('identifier - unmodified', t => {
   t.is(toIdentifier('abc'), 'abc');
@@ -33,38 +33,40 @@ test('identifier - remove starting numbers', t => {
 });
 
 test('identifier - empty string', t => {
-  let error: OptionsError | undefined = t.throws(() => toIdentifier(''));
-  t.is(error?.messages.name, 'Identifier is empty or does not have valid characters');
+  let error = t.throws(() => toIdentifier(''), { instanceOf: OptionsError });
+  t.is(error.messages.name, 'Identifier is empty or does not have valid characters');
 });
 
 test('identifier - no valid chars', t => {
-  let error: OptionsError | undefined = t.throws(() => toIdentifier('123'));
-  t.is(error?.messages.name, 'Identifier is empty or does not have valid characters');
+  let error = t.throws(() => toIdentifier('123'),  { instanceOf: OptionsError });
+  t.is(error.messages.name, 'Identifier is empty or does not have valid characters');
 });
 
-test('short string - unmodified', t => {
-  t.is(toShortString('abc', 'foo'), 'abc');
+test('toStringLiteral - unmodified', t => {
+  t.is(toStringLiteral('abc'), 'abc');
 });
 
-test('short string - remove accents', t => {
-  t.is(toShortString('ábc', 'foo'), 'abc');
+test('toStringLiteral - remove accents', t => {
+  t.is(toStringLiteral('ábc'), 'abc');
 });
 
-test('short string - remove non-ascii-printable characters', t => {
-  t.is(toShortString('abc😀', 'foo'), 'abc');
+test('toStringLiteral - remove non-ascii-printable characters', t => {
+  t.is(toStringLiteral('abc😀'), 'abc');
 });
 
-test('short string - escape single quote', t => {
-  t.is(toShortString("abc'def", 'foo'), "abc\\'def");
+test('toStringLiteral - escape double quote', t => {
+  t.is(toStringLiteral("abc\"def"), "abc\\\"def");
 });
 
-test('short string - escape backslash', t => {
-  t.is(toShortString('abc\\def', 'foo'), 'abc\\\\def');
+test('toStringLiteral - does not escape single quote', t => {
+  t.is(toStringLiteral("abc'def"), "abc'def");
 });
 
-test('short string - max 31 characters', t => {
-  t.is(toShortString('A234567890123456789012345678901', 'foo'), 'A234567890123456789012345678901');
+test('toStringLiteral - escape backslash', t => {
+  t.is(toStringLiteral('abc\\def'), 'abc\\\\def');
+});
 
-  let error: OptionsError | undefined = t.throws(() => toShortString('A2345678901234567890123456789012', 'foo'));
-  t.is(error?.messages.foo, 'String is longer than 31 characters');
+test('more than 31 characters', t => {
+  t.is(toStringLiteral('A234567890123456789012345678901'), 'A234567890123456789012345678901');
+  t.is(toStringLiteral('A2345678901234567890123456789012'), 'A2345678901234567890123456789012');
 });