+++ packages/core/ava.config.js
@@ -1,11 +1,13 @@
 module.exports = {
   extensions: ['ts'],
   require: ['ts-node/register'],
-  ignoredByWatcher: [
-    'contracts',
-    'artifacts',
-    'cache',
-  ],
+  watchmode: {
+    ignoreChanges: [
+      'contracts',
+      'artifacts',
+      'cache',
+    ],
+  },
   timeout: '10m',
   workerThreads: false,
 };