+++ packages/core/src/generate/erc20.ts
@@ -1,5 +1,6 @@
 import type { ERC20Options } from '../erc20';
 import { accessOptions } from '../set-access-control';
+import { clockModeOptions } from '../set-clock-mode';
 import { infoOptions } from '../set-info';
 import { upgradeableOptions } from '../set-upgradeable';
 import { generateAlternatives } from './alternatives';
@@ -13,7 +14,7 @@ const blueprint = {
   pausable: booleans,
   mintable: booleans,
   permit: booleans,
-  votes: booleans,
+  votes: [ ...booleans, ...clockModeOptions ] as const,
   flashmint: booleans,
   premint: ['1'],
   access: accessOptions,