+++ packages/core/src/index.ts
@@ -20,3 +20,5 @@ export type { Kind } from './kind';
 export { sanitizeKind } from './kind';
 
 export { erc20, erc721, erc1155, governor, custom } from './api';
+
+export { compatibleContractsSemver } from './utils/version';