+++ packages/ui/src/cairo/ERC1155Controls.svelte
@@ -1,42 +1,33 @@
 <script lang="ts">
   import HelpTooltip from '../HelpTooltip.svelte';
 
-  import type { KindedOptions, OptionsErrorMessages } from '@openzeppelin/wizard-cairo';
-  import { erc721, infoDefaults } from '@openzeppelin/wizard-cairo';
-  
+  import type { KindedOptions } from '@openzeppelin/wizard-cairo';
+  import { erc1155, infoDefaults } from '@openzeppelin/wizard-cairo';
+
   import AccessControlSection from './AccessControlSection.svelte';
   import UpgradeabilityField from './UpgradeabilityField.svelte';
   import InfoSection from './InfoSection.svelte';
-  import { error } from '../error-tooltip';
-  
-  export const opts: Required<KindedOptions['ERC721']> = {
-    kind: 'ERC721',
-    ...erc721.defaults,
+
+  export let opts: Required<KindedOptions['ERC1155']> = {
+    kind: 'ERC1155',
+    ...erc1155.defaults,
     info: { ...infoDefaults }, // create new object since Info is nested
   };
 
-  export let errors: undefined | OptionsErrorMessages;
-
-  $: requireAccessControl = erc721.isAccessControlRequired(opts);
+  $: requireAccessControl = erc1155.isAccessControlRequired(opts);
 </script>
 
 <section class="controls-section">
   <h1>Settings</h1>
 
-  <div class="grid grid-cols-[2fr,1fr] gap-2">
-    <label class="labeled-input">
-      <span>Name</span>
-      <input bind:value={opts.name} use:error={errors?.name}>
-    </label>
-    <label class="labeled-input">
-      <span>Symbol</span>
-      <input bind:value={opts.symbol} use:error={errors?.symbol}>
-    </label>
-  </div>
+  <label class="labeled-input">
+    <span>Name</span>
+    <input bind:value={opts.name}>
+  </label>
   <label class="labeled-input">
     <span class="flex justify-between pr-2">
       Base URI
-      <HelpTooltip>Will be concatenated with token IDs to generate the token URIs.</HelpTooltip>
+      <HelpTooltip>Location of the metadata. Clients will replace any instance of {"{id}"} in this string with the tokenId.</HelpTooltip>
     </span>
     <input bind:value={opts.baseUri} placeholder="https://...">
   </label>
@@ -49,8 +40,8 @@
     <label class:checked={opts.mintable}>
       <input type="checkbox" bind:checked={opts.mintable}>
       Mintable
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/erc721">
-        Privileged accounts will be able to emit new tokens.
+      <HelpTooltip>
+        Privileged accounts will be able to create more supply.
       </HelpTooltip>
     </label>
     <label class:checked={opts.burnable}>
@@ -66,6 +57,13 @@
       <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/security#pausable">
         Privileged accounts will be able to pause the functionality marked with <code>self.pausable.assert_not_paused()</code>.
         Useful for emergency response.
+      </HelpTooltip>
+    </label>
+    <label class:checked={opts.updatableUri}>
+      <input type="checkbox" bind:checked={opts.updatableUri}>
+      Updatable URI
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/api/erc1155#ERC1155Component-set_base_uri">
+        Privileged accounts will be able to set a new URI for all token types. Clients will replace any instance of {"{id}"} in the URI with the tokenId.
       </HelpTooltip>
     </label>
     <UpgradeabilityField bind:upgradeable={opts.upgradeable} />