+++ packages/ui/src/main.ts
@@ -4,6 +4,10 @@ import type {} from 'svelte';
 import App from './App.svelte';
 import CairoApp from './cairo/App.svelte';
 import { postMessage } from './post-message';
+import UnsupportedVersion from './UnsupportedVersion.svelte';
+import semver from 'semver';
+import { compatibleContractsSemver as compatibleSolidityContractsSemver } from '@openzeppelin/wizard';
+import { compatibleContractsSemver as compatibleCairoContractsSemver } from '@openzeppelin/wizard-cairo';
 
 function postResize() {
   const { height } = document.documentElement.getBoundingClientRect();
@@ -18,10 +22,16 @@ resizeObserver.observe(document.body);
 const params = new URLSearchParams(window.location.search);
 
 const initialTab = params.get('tab') ?? undefined;
-const lang = params.get('lang');
+const lang = params.get('lang') ?? undefined;
+const requestedVersion = params.get('version') ?? undefined;
+
+let compatibleVersionSemver = lang === 'cairo' ? compatibleCairoContractsSemver : compatibleSolidityContractsSemver;
 
 let app;
-if (lang === 'cairo') {
+if (requestedVersion && !semver.satisfies(requestedVersion, compatibleVersionSemver)) {
+  postMessage({ kind: 'oz-wizard-unsupported-version' });
+  app = new UnsupportedVersion({ target: document.body, props: { requestedVersion, compatibleVersionSemver }});
+} else if (lang === 'cairo') {
   app = new CairoApp({ target: document.body, props: { initialTab } });
 } else {
   app = new App({ target: document.body, props: { initialTab } });